farmer/requirements.txt
face_recognition==1.3.0
numpy==1.26.3
opencv_python==4.9.0.80
pandas==2.2.0
PyAutoGUI==0.9.54
Pillow==10.2.0
pyarrow==15.0.0